+++ HISTORY.rst
@@ -20,7 +20,7 @@ latest (unreleased)
 +++++++++++++++++++
 
 **Features and Improvements**
-
+* BIZ-1225: Script to clean salary structure set on employee and clean dupplicate structure
 * BIZ-1234: Base de salaire ne pas adapter à la société courante
 
 **Bugfixes**